Soft deletes (#139)
* feat(api): soft deletion for gatsby incremental builds

* style: fix StyleCI findings

+++ app/Listeners/Image/RemoveImageFromStorage.php
@@ -21,6 +21,8 @@ class RemoveImageFromStorage implements ShouldQueue
     {
         $image = $event->getImage();
 
-        Storage::disk('images')->delete($image->path);
+        if ($image->isForceDeleting()) {
+            Storage::disk('images')->delete($image->path);
+        }
     }
 }
